+++ src/components/App.tsx
@@ -167,6 +167,7 @@ import {
   NonDeletedExcalidrawElement,
   ExcalidrawTextContainer,
   ExcalidrawFrameElement,
+  ExcalidrawIFrameElement,
 } from "../element/types";
 import { getCenter, getDistance } from "../gesture";
 import {
@@ -256,8 +257,9 @@ import {
   isTransparent,
   easeToValuesRAF,
   muteFSAbortError,
+  easeOut,
 } from "../utils";
-import { IFrameURLValidator, getEmbedLink } from "../element/iframe";
+import { iframeURLValidator, getEmbedLink } from "../element/iframe";
 import {
   ContextMenu,
   ContextMenuItems,
@@ -304,14 +306,13 @@ import {
 import { isHittingElementNotConsideringBoundingBox } from "../element/collision";
 import { resizeSingleElement } from "../element/resizeElements";
 import {
-  normalizeLink,
   showHyperlinkTooltip,
   hideHyperlinkToolip,
   Hyperlink,
   isPointHittingLink,
-  isLocalLink,
   isPointHittingLinkIcon,
 } from "../element/Hyperlink";
+import { isLocalLink, normalizeLink } from "../data/url";
 import { ImportedDataState } from "../data/types"; //zsviczian
 import { shouldShowBoundingBox } from "../element/transformHandles";
 import { actionUnlockAllElements } from "../actions/actionElementLock";
@@ -336,10 +337,7 @@ import {
   actionRemoveAllElementsFromFrame,
   actionSelectAllElementsInFrame,
 } from "../actions/actionFrame";
-import {
-  actionToggleHandTool,
-  zoomToFitElements,
-} from "../actions/actionCanvas";
+import { actionToggleHandTool, zoomToFit } from "../actions/actionCanvas";
 import { jotaiStore } from "../jotai";
 import { activeConfirmDialogAtom } from "./ActiveConfirmDialog";
 import { actionWrapTextInContainer } from "../actions/actionBoundText";
@@ -499,7 +497,6 @@ class App extends React.Component<AppProps, AppState> {
       ...(initState ?? {}), //zsviczian
       defaultSidebarDockedPreference: false,
     };
-    IFrameURLValidator.getInstance(props.iframeURLWhitelist);
     this.id = nanoid();
     this.library = new Library(this);
     if (excalidrawRef) {
@@ -795,17 +792,37 @@ class App extends React.Component<AppProps, AppState> {
     );
   }
 
+  previsloading = true;
   private renderIFrames() {
     const scale = this.state.zoom.value;
     const normalizedWidth = this.state.width;
     const normalizedHeight = this.state.height;
-    const iFrameElements = this.scene
-      .getNonDeletedElements()
-      .filter((el) => isIFrameElement(el) && el.whitelisted);
+    const whitelistUpdates = new Map<ExcalidrawIFrameElement, boolean>();
+
+    const iFrameElements = this.scene.getNonDeletedElements().filter((el) => {
+      if (!isIFrameElement(el)) {
+        return false;
+      }
+      if (typeof el.whitelisted === "undefined") {
+        const isWhitelisted = iframeURLValidator(
+          el.link,
+          this.props.iframeURLWhitelist,
+        );
+        whitelistUpdates.set(el, isWhitelisted);
+        return isWhitelisted;
+      }
+      return el.whitelisted;
+    });
     Object.keys(this.iFrameRefs).forEach((key) => {
       if (!iFrameElements.some((el) => el.id === key)) {
         delete this.iFrameRefs[key];
       }
+    });
+    setTimeout(() => {
+      whitelistUpdates.forEach((whitelisted, element) => {
+        mutateElement(element, { whitelisted });
+        invalidateShapeForElement(element);
+      });
     });
     return (
       <>
@@ -847,6 +864,22 @@ class App extends React.Component<AppProps, AppState> {
               }}
             >
               <div
+                //this is a hack that addresses isse with embedded excalidraw.com iframe
+                //https://github.com/excalidraw/excalidraw/pull/6691#issuecomment-1607383938
+                /*ref={(ref) => {
+                  if (!this.excalidrawContainerRef.current) {
+                    return;
+                  }
+                  const container = this.excalidrawContainerRef.current;
+                  const sh = container.scrollHeight;
+                  const ch = container.clientHeight;
+                  if (sh !== ch) {
+                    container.style.height = `${sh}px`;
+                    setTimeout(() => {
+                      container.style.height = `100%`;
+                    });
+                  }
+                }}*/
                 className="excalidraw__iframe-container__inner"
                 style={{
                   width: isVisible ? `${el.width}px` : 0,
@@ -856,10 +889,16 @@ class App extends React.Component<AppProps, AppState> {
                   pointerEvents: isActive ? "auto" : "none",
                 }}
               >
+                {isHovered && (
+                  <div className="excalidraw__iframe-hint">
+                    {t("buttons.iframeInteractionButton")}
+                  </div>
+                )}
                 <div
                   style={{
                     padding: `${el.strokeWidth}px`,
                     height: "100%",
+                    width: "100%",
                   }}
                 >
                   {this.props.renderCustomIFrame?.(
@@ -2236,7 +2275,7 @@ class App extends React.Component<AppProps, AppState> {
       } else if (data.text) {
         if (
           !isPlainPaste &&
-          IFrameURLValidator.getInstance().run(data.text) &&
+          iframeURLValidator(data.text, this.props.iframeURLWhitelist) &&
           (/^(http|https):\/\/[^\s/$.?#].[^\s]*$/.test(data.text) ||
             getEmbedLink(data.text)?.type === "video")
         ) {
@@ -2562,27 +2601,51 @@ class App extends React.Component<AppProps, AppState> {
     target:
       | ExcalidrawElement
       | readonly ExcalidrawElement[] = this.scene.getNonDeletedElements(),
-    opts?: { fitToContent?: boolean; animate?: boolean; duration?: number },
+    opts?:
+      | {
+          fitToContent?: boolean;
+          fitToViewport?: never;
+          viewportZoomFactor?: never;
+          animate?: boolean;
+          duration?: number;
+        }
+      | {
+          fitToContent?: never;
+          fitToViewport?: boolean;
+          /** when fitToViewport=true, how much screen should the content cover,
+           * between 0.1 (10%) and 1 (100%)
+           */
+          viewportZoomFactor?: number;
+          animate?: boolean;
+          duration?: number;
+        },
   ) => {
     this.cancelInProgresAnimation?.();
 
     // convert provided target into ExcalidrawElement[] if necessary
-    const targets = Array.isArray(target) ? target : [target];
+    const targetElements = Array.isArray(target) ? target : [target];
 
     let zoom = this.state.zoom;
     let scrollX = this.state.scrollX;
     let scrollY = this.state.scrollY;
 
-    if (opts?.fitToContent) {
-      // compute an appropriate viewport location (scroll X, Y) and zoom level
-      // that fit the target elements on the scene
-      const { appState } = zoomToFitElements(targets, this.state, false);
+    if (opts?.fitToContent || opts?.fitToViewport) {
+      const { appState } = zoomToFit({
+        targetElements,
+        appState: this.state,
+        fitToViewport: !!opts?.fitToViewport,
+        viewportZoomFactor: opts?.viewportZoomFactor,
+      });
       zoom = appState.zoom;
       scrollX = appState.scrollX;
       scrollY = appState.scrollY;
     } else {
       // compute only the viewport location, without any zoom adjustment
-      const scroll = calculateScrollCenter(targets, this.state, this.canvas);
+      const scroll = calculateScrollCenter(
+        targetElements,
+        this.state,
+        this.canvas,
+      );
       scrollX = scroll.scrollX;
       scrollY = scroll.scrollY;
     }
@@ -2592,19 +2655,42 @@ class App extends React.Component<AppProps, AppState> {
     if (opts?.animate) {
       const origScrollX = this.state.scrollX;
       const origScrollY = this.state.scrollY;
+      const origZoom = this.state.zoom.value;
 
-      // zoom animation could become problematic on scenes with large number
-      // of elements, setting it to its final value to improve user experience.
-      //
-      // using zoomCanvas() to zoom on current viewport center
-      this.zoomCanvas(zoom.value);
+      const cancel = easeToValuesRAF({
+        fromValues: {
+          scrollX: origScrollX,
+          scrollY: origScrollY,
+          zoom: origZoom,
+        },
+        toValues: { scrollX, scrollY, zoom: zoom.value },
+        interpolateValue: (from, to, progress, key) => {
+          // for zoom, use different easing
+          if (key === "zoom") {
+            return from * Math.pow(to / from, easeOut(progress));
+          }
+          // handle using default
+          return undefined;
+        },
+        onStep: ({ scrollX, scrollY, zoom }) => {
+          this.setState({
+            scrollX,
+            scrollY,
+            zoom: { value: zoom },
+          });
+        },
+        onStart: () => {
+          this.setState({ shouldCacheIgnoreZoom: true });
+        },
+        onEnd: () => {
+          this.setState({ shouldCacheIgnoreZoom: false });
+        },
+        onCancel: () => {
+          this.setState({ shouldCacheIgnoreZoom: false });
+        },
+        duration: opts?.duration ?? 500,
+      });
 
-      const cancel = easeToValuesRAF(
-        [origScrollX, origScrollY],
-        [scrollX, scrollY],
-        (scrollX, scrollY) => this.setState({ scrollX, scrollY }),
-        { duration: opts?.duration ?? 500 },
-      );
       this.cancelInProgresAnimation = () => {
         cancel();
         this.cancelInProgresAnimation = null;
@@ -3886,12 +3972,19 @@ class App extends React.Component<AppProps, AppState> {
       this.device.isMobile,
     );
     if (lastPointerDownHittingLinkIcon && lastPointerUpHittingLinkIcon) {
-      const url = this.hitLinkElement.link;
+      let url = this.hitLinkElement.link;
       if (url) {
+        url = normalizeLink(url);
         let customEvent;
         if (this.props.onLinkOpen) {
           customEvent = wrapEvent(EVENT.EXCALIDRAW_LINK, event.nativeEvent);
-          this.props.onLinkOpen(this.hitLinkElement, customEvent);
+          this.props.onLinkOpen(
+            {
+              ...this.hitLinkElement,
+              link: url,
+            },
+            customEvent,
+          );
         }
         if (!customEvent?.defaultPrevented) {
           const target = isLocalLink(url) ? "_self" : "_blank";
@@ -3899,7 +3992,7 @@ class App extends React.Component<AppProps, AppState> {
           // https://mathiasbynens.github.io/rel-noopener/
           if (newWindow) {
             newWindow.opener = null;
-            newWindow.location = normalizeLink(url);
+            newWindow.location = url;
           }
         }
       }
@@ -5465,7 +5558,7 @@ class App extends React.Component<AppProps, AppState> {
       width: embedLink.aspectRatio.w,
       height: embedLink.aspectRatio.h,
       link,
-      whitelisted: IFrameURLValidator.getInstance().run(link),
+      whitelisted: undefined,
     });
 
     this.scene.replaceAllElements([
@@ -7604,7 +7697,7 @@ class App extends React.Component<AppProps, AppState> {
       const text = event.dataTransfer?.getData("text");
       if (
         text &&
-        IFrameURLValidator.getInstance().run(text) &&
+        iframeURLValidator(text, this.props.iframeURLWhitelist) &&
         (/^(http|https):\/\/[^\s/$.?#].[^\s]*$/.test(text) ||
           getEmbedLink(text)?.type === "video")
       ) {