+++ src/data/restore.ts
@@ -41,7 +41,7 @@ import {
   measureBaseline,
 } from "../element/textElement";
 import { COLOR_PALETTE } from "../colors";
-import { IFrameURLValidator } from "../element/iframe";
+import { normalizeLink } from "./url";
 
 type RestoredAppState = Omit<
   AppState,
@@ -144,7 +144,7 @@ const restoreElementWithProperties = <
       ? element.boundElementIds.map((id) => ({ type: "arrow", id }))
       : element.boundElements ?? [],
     updated: element.updated ?? getUpdatedTimestamp(),
-    link: element.link ?? null,
+    link: element.link ? normalizeLink(element.link) : null,
     locked: element.locked ?? false,
   };
 
@@ -279,7 +279,7 @@ const restoreElement = (
       return restoreElementWithProperties(element, {});
     case "iframe":
       return restoreElementWithProperties(element, {
-        whitelisted: IFrameURLValidator.getInstance().run(element.link),
+        whitelisted: undefined,
       });
     case "frame":
       return restoreElementWithProperties(element, {