+++ src/types.ts
@@ -411,7 +411,7 @@ export interface ExcalidrawProps {
   ) => void;
   onScrollChange?: (scrollX: number, scrollY: number) => void;
   children?: React.ReactNode;
-  iframeURLWhitelist?: RegExp[];
+  iframeURLWhitelist?: IFrameURLValidatorRules;
   renderCustomIFrame?: (
     element: NonDeletedExcalidrawElement,
     radius: number,
@@ -644,3 +644,5 @@ export type FrameNameBoundsCache = {
     }
   >;
 };
+
+export type IFrameURLValidatorRules = (RegExp | boolean)[];