+++ src/packages/excalidraw/CHANGELOG.md
@@ -15,7 +15,7 @@ Please add the latest change on the top under the correct section.
 
 ### iFrame Site Whitelist
 
-- Set `props.iframeURLWhitelist` to allow additional sites to be added as iFrames. The default list includes: YouTube, Vimeo, Figma, Excalidraw, Twitter. The prop takes a `RegExp[]` array. Set the prop to `[/.*/]` if you want to allow all sites.
+- Set `props.iframeURLWhitelist` to allow additional sites to be added as iFrames. The default list includes: YouTube, Vimeo, Figma. The prop takes a `(RegExp | boolean)[]` array. If you pass `[true]` all URLs will be allowed, if you pass `[...,false]`, then the rest of the validators after `false` will not be executed, including the omission of the excalidraw default validators. If you only send `RegExp[]` then after running the validators Excalidraw will also execute the built in validators.
 
 ### renderCustomIFrame
 
@@ -37,6 +37,8 @@ JSX.Element | null: The JSX component representing the custom rendering, or null
 
 ### Features
 
+- Add support for `opts.fitToViewport` and `opts.viewportZoomFactor` in the [`ExcalidrawAPI.scrollToContent`](https://docs.excalidraw.com/docs/@excalidraw/excalidraw/api/props/ref#scrolltocontent) API. [#6581](https://github.com/excalidraw/excalidraw/pull/6581).
+- Properly sanitize element `link` urls. [#6728](https://github.com/excalidraw/excalidraw/pull/6728).
 - Sidebar component now supports tabs — for more detailed description of new behavior and breaking changes, see the linked PR. [#6213](https://github.com/excalidraw/excalidraw/pull/6213)
 - Exposed `DefaultSidebar` component to allow modifying the default sidebar, such as adding custom tabs to it. [#6213](https://github.com/excalidraw/excalidraw/pull/6213)
 
@@ -85,7 +87,7 @@ JSX.Element | null: The JSX component representing the custom rendering, or null
 
 ### Features
 
-- [`ExcalidrawAPI.scrolToContent`](https://docs.excalidraw.com/docs/@excalidraw/excalidraw/api/props/ref#scrolltocontent) has new opts object allowing you to fit viewport to content, and animate the scrolling. [#6319](https://github.com/excalidraw/excalidraw/pull/6319)
+- [`ExcalidrawAPI.scrollToContent`](https://docs.excalidraw.com/docs/@excalidraw/excalidraw/api/props/ref#scrolltocontent) has new opts object allowing you to fit viewport to content, and animate the scrolling. [#6319](https://github.com/excalidraw/excalidraw/pull/6319)
 
 - Expose `useI18n()` hook return an object containing `t()` i18n helper and current `langCode`. You can use this in components you render as `<Excalidraw>` children to render any of our i18n locale strings. [#6224](https://github.com/excalidraw/excalidraw/pull/6224)